Added xattr support for quering the cached status of a file in the UG. Query the syndicate_cached_blocks attribute, and receive a bit string where the ith bit is 1 if block i in the file is cached.  The getxattr() syscall for this particular attribute scans the cached data on-disk, so it remains accurate in the face of user-evicted cached blocks.  This commit also adds support for the syndicate_coordinator attribute, which when queried with getxattr() will return the name of the UG that serves as the write coordinator.
This commit additionally includes a few
bug-fixes discovered along the way, as well as a re-factoring of the
xattrs.cpp file.  You can set, list and delete attributes in the UG, but
they will not (yet) preserved across instances or across the volume.  In
the near future, setxattr() and removexattr() will synchronously send
the new attributes to the MS, so other UGs can read them.