Introduce the GC log to the MS, which will track all prior manifest modtimes (i.e. modification nonces) for each file. This will let the coordinator UG query the prior manifest modtimes, allowing it to obtain un-garbage-collected manifests and blocks from the RG.  This will allow it to asynchronously garbage-collect older versions of a file that are no longer needed, even in the face of partial or perminent failures of individual UGs.  So long as the UG can become a file's coordinator, it can garbage-collect its old data.
WIP: To ensure that a rogue UG can't purge this log, it will present "deletion receipts" from the RGs, which indicate that
the RG has certified that the manifest with the given timestamp has been deleted.  The MS will only remove a GC log entry
if all of the Volume's current RGs generated deletion receipts for the associated manifest (which the UG will present to
the MS in the request).